Refactored for easier testing
The current code depdended completely on the platform as it performed
live tests in the unit testing code. This patch refactors the
test code to create a mock `ldd` script which is executed instead of
the actual script.

A few minor bugs were uncovered and were fixed (brackets around
the memory address and special case of VDSO).